Added of "statementType" to batch result for ODBC
- new statementType field in results so that ODBC driver can set the status field appropriately
  (still to be properly implemented on the server side)
- add explicit std::move in a couple of places when returning from function

+++ native/src/snappyclient/overrides/snappydata_struct_ColumnValue.h
@@ -292,7 +292,7 @@ public:
   // Similarly when getting column values as result, we could send a char*
   // buffer to read the result into instead of creating a new std::string
   // (though it is dangerous and should avoid unless perf benefit is proven)
-  inline void setString(const char* val, const int32_t len) {
+  inline void setString(const char* val, const size_t len) {
     if (len < 0) {
       m_val = std::move(std::shared_ptr<std::string>(new std::string(val)));
     } else {
